+++ brewtils/models.py
@@ -241,6 +241,7 @@ class Instance(BaseModel):
         "STARTING",
         "STOPPING",
         "UNKNOWN",
+        "AWAITING_SYSTEM",
     }
 
     def __init__(
@@ -821,6 +822,8 @@ class System(BaseModel):
         template=None,
         groups=None,
         prefix_topic=None,
+        requires=None,
+        requires_timeout=None,
     ):
         self.name = name
         self.description = description
@@ -837,6 +840,8 @@ class System(BaseModel):
         self.template = template
         self.groups = groups or []
         self.prefix_topic = prefix_topic
+        self.requires = requires or []
+        self.requires_timeout = requires_timeout
 
     def __str__(self):
         return "%s:%s-%s" % (self.namespace, self.name, self.version)