+++ brewtils/test/fixtures.py
@@ -270,6 +270,8 @@ def system_dict(instance_dict, command_dict, command_dict_2, system_id):
         "template": "<html>template</html>",
         "groups": ["GroupB", "GroupA"],
         "prefix_topic": "custom_topic",
+        "requires": ["SystemA"],
+        "requires_timeout": 300,
     }